Refactoring: move npm command to NPM context
