--- src/main/java/tech/jhipster/lite/generator/project/domain/CommandRepository.java
@@ -1,6 +1,0 @@
-package tech.jhipster.lite.generator.project.domain;
-
-public interface CommandRepository {
-  void npmInstall(Project project);
-  void npmPrettierFormat(Project project);
-}
